+++ src/library/DIPS.Mobile.UI/Components/Buttons/Button.cs
@@ -1,7 +1,3 @@
-using DIPS.Mobile.UI.Resources.Colors;
-using DIPS.Mobile.UI.Extensions;
-using DIPS.Mobile.UI.Sizes.Sizes;
-
 namespace DIPS.Mobile.UI.Components.Buttons
 {
     public partial class Button : Microsoft.Maui.Controls.Button
@@ -10,8 +6,8 @@ namespace DIPS.Mobile.UI.Components.Buttons
         {
             this.SetAppThemeColor(BackgroundColorProperty, ColorName.color_primary_90);
             this.SetAppThemeColor(TextColorProperty, ColorName.color_neutral_05);
-            Padding = new Thickness(UI.Resources.Sizes.Sizes.GetSize(SizeName.size_3), UI.Resources.Sizes.Sizes.GetSize(SizeName.size_1));
-            CornerRadius = UI.Resources.Sizes.Sizes.GetSize(SizeName.size_2);
+            Padding = new Thickness(Sizes.GetSize(SizeName.size_3), Sizes.GetSize(SizeName.size_1));
+            CornerRadius = Sizes.GetSize(SizeName.size_2);
         }