+++ src/library/DIPS.Mobile.UI/Components/Lists/VerticalStackLayout.cs
@@ -1,4 +1,4 @@
-using DIPS.Mobile.UI.Sizes.Sizes;
+using DIPS.Mobile.UI.Resources.Sizes;
 
 namespace DIPS.Mobile.UI.Components.Lists;
 
@@ -6,6 +6,6 @@ public class VerticalStackLayout : Microsoft.Maui.Controls.VerticalStackLayout
 {
     public VerticalStackLayout()
     {
-        Spacing = UI.Resources.Sizes.Sizes.GetSize(SizeName.size_1);
+        Spacing = Sizes.GetSize(SizeName.size_1);
     }
 }