+++ src/library/DIPS.Mobile.UI/GlobalUsings.cs
@@ -1,0 +1,6 @@
+global  using DIPS.Mobile.UI;
+global using DIPS.Mobile.UI.Extensions;
+global using DIPS.Mobile.UI.Resources.Sizes;
+global using DIPS.Mobile.UI.Resources.Colors;
+global using DIPS.Mobile.UI.Resources.Icons;
+global using DIPS.Mobile.UI.Resources.LocalizedStrings;