+++ src/library/DIPS.Mobile.UI/Resources/Sizes/Sizes.xaml.cs
@@ -1,5 +1,3 @@
-using DIPS.Mobile.UI.Sizes.Sizes;
-
 namespace DIPS.Mobile.UI.Resources.Sizes;
 
 public partial class Sizes